gke/fault-trigger-ui/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.7.1
